Add CIChangeCMDBHistory into SplunkLog and refactoring SplunkLogger
