+++ src/ralph/cmdb/integration/splunk.py
@@ -9,6 +9,7 @@ import socket
 
 from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
+from django.core.urlresolvers import reverse
 
 SPLUNK_PORT = settings.SPLUNK_LOGGER_PORT
 SPLUNK_HOST = settings.SPLUNK_LOGGER_HOST
@@ -40,5 +41,22 @@ class SplunkLogger(object):
         return ' '.join(
             ['|#{}={}#|'.format(
                 repr(key), repr(value)
-            ) for (key, value) in message.items()]
-        ).replace("u", '')
+            ) for (key, value) in message.items() if not key.startswith('_')]
+        )
+
+
+def log_change_to_splunk(instance, log_type):
+    message = vars(instance)
+    message['type'] = log_type
+    if hasattr(instance, 'ci'):
+        message['ci_name'] = instance.ci.name if instance.ci else None
+        message['ralph_link'] = reverse(
+            'ci_view_main', kwargs={'ci_id': instance.ci.id}
+        ) if instance.ci else None
+    if hasattr(instance, 'user'):
+        message['author'] = instance.user.username
+    if hasattr(instance, 'device'):
+        message['device_name'] = instance.device.name
+        message['venture'] = instance.device.venture.name
+        message['role'] = instance.device.venture_role.name
+    SplunkLogger().log_dict(message)
